update Cooking Skill adapter for recent versions of Cooking Skill and SMAPI 3.0
