+++ SkillPrestige.CookingSkill/manifest.json
@@ -1,22 +1,13 @@
 {
-  "Name": "Cooking Skill Prestige Adapter",
+  "Name": "Skill Prestige: Cooking Skill Adapter",
   "Author": "Alphablackwolf",
-  "Version": {
-    "MajorVersion": 1,
-    "MinorVersion": 1,
-    "PatchVersion": 0,
-    "Build": null
-  },
-  "Description": "",
-  "MinimumApiVersion": "1.14",
+  "Version": "1.1.0",
+  "Description": "Adds Cooking Skill support to Skill Prestige.",
   "UniqueID": "Alphablackwolf.CookingSkillPrestigeAdapter",
-  "EntryDll": "SkillPrestigeCookingAdapter.dll",
+  "EntryDll": "SkillPrestige.CookingSkill.dll",
+  "MinimumApiVersion": "2.11.0",
   "Dependencies": [
-    {
-      "UniqueID": "spacechase0.CookingSkill"
-    },
-    {
-      "UniqueID": "alphablackwolf.skillPrestige"
-    }
+    { "UniqueID": "alphablackwolf.skillPrestige" },
+    { "UniqueID": "spacechase0.CookingSkill" }
   ]
 }
