+++ SkillPrestige.LuckSkill/manifest.json
@@ -3,15 +3,11 @@
   "Author": "Alphablackwolf",
   "Version": "1.2.3",
   "Description": "Adds Luck Skill support to Skill Prestige.",
-  "UniqueID": "alphablackwolf.skillPrestigeLuckSkill",
+  "UniqueID": "alphablackwolf.LuckSkillPrestigeAdapter",
   "EntryDll": "SkillPrestige.LuckSkill.dll",
-  "MinimumApiVersion": "2.10.2",
+  "MinimumApiVersion": "2.11.0",
   "Dependencies": [
-    {
-      "UniqueID": "alphablackwolf.skillPrestige"
-    },
-    {
-      "UniqueID": "spacechase0.LuckSkill"
-    }
+    { "UniqueID": "alphablackwolf.skillPrestige" },
+    { "UniqueID": "spacechase0.LuckSkill" }
   ]
 }
